+++ CHANGELOG.md
@@ -34,7 +34,7 @@
   ],
   ...
   ```
-- Adds support for plugin views.
+- Adds support for plugin view types and plugin coordination types.
 
 ### Changed
 - Merged dependabot PRs.